Android: Fix compile errors
Android doesn't have DBUS so all functionality related to it
must be disabled by checking if GLIB_SUPPORT is defined.

Currently, the build also fails with the following error:
thd_trip_point.h:139:2: error: control may reach end of non-void function

This is because the exception is excluded for Android, and nothing is
returned if an invalid index is given into the method.

The easiest way to fix this is to exclude the method entirely for
Android. The method is only used from DBUS code so it doesn't
cause any problems if Android doesn't have it.