Use pid output offset from the min_state
Instead of using current state and adjust the next state relative to
this based on output use min_state, which is fixed. The curr_state
always change which causes difficult to control.
Why not direct use PID output?
The problem is that we have devices where min_state > max_state. In
this case output written to device needs to be smaller for more
cooling action. In this case the Kp can be negative value, but
that won't be enough. By subtracting the PID output from min_state
will result in control value which is less than min_state.