Honor ACPI _CRT for processor thermal zone
There are some new fanless platforms use DPTF's virtual sensor instead
of INT340X devices.

Because of that, the _PSV is no longer in use, at least not directly,
hence its value may set higher then _CRT. To a fanless system that means
no cooling device gets activated before _CRT, so the system will be
considered overheated by Linux kernel, and gets shutdown by the kernel.

To avoid that, we use _CRT - 15C as our new _PSV value.
A 15C margin below _CRT should be enough for passive cooling devices to
clamp the temperature down.

This is a compromise, but we need this until we have full DPTF support.